+++ _pytest/test_sanity.py
@@ -34,6 +34,7 @@ def import_submodules(package_name, skip_list):
     return results
 
 
+@pytest.mark.parallel
 def test_no_global_imports_of_banned_package():
     """This test ensures that neither of the banned packages are imported module wise in any of our code files.